Implementacion de funcionalidad boton buscar en gestion
Se implemento la funcionalidad del boton buscar que ofrecer toda la informacion del cliente que se busca por dni.

+++ src/main/java/com/utp/web/TrabajoFinalWEB/models/entity/Inscripcion.java
@@ -47,7 +47,7 @@ public class Inscripcion implements Serializable {
     private String estado;
 
     @Column(nullable = false)
-    private Long diasInscripcion;
+    private Integer diasInscripcion;
 
     
 }
